requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
python-dotenv==1.0.1
APScheduler==3.10.4
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
click==8.1.7
paho-mqtt==2.1.0
prometheus-client==0.21.0
requests==2.32.3

